pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "monipy"
version = "0.1.0"
description = "wearable seizure detection"
requires-python = ">=3.8,<3.9"

dependencies = [
  "numpy==1.23.5",
  "scipy>=1.10,<1.11",
  "pandas>=2.0,<2.1",
  "scikit-learn>=1.3,<1.4",
  "matplotlib>=3.7,<3.8",
  "seaborn>=0.13,<0.14",
  "mlflow>=2.17,<2.18",
  "optuna>=4.5,<4.6",
  "shap>=0.44,<0.45",
  "pyarrow>=17,<18",
  "torch==2.2.2",
  "torchvision==0.17.2",
  "torchaudio==2.2.2",
  "fastai>=2.7,<2.8",
  "tsai>=0.3,<0.4",
  "biosppy>=2.2,<2.3",
  "neurokit2>=0.2,<0.3",
  "wfdb>=4.1,<4.2",
  "py-ecg-detectors>=1.3,<1.4",
  "PeakUtils>=1.3,<1.4",
  "PyWavelets>=1.4,<1.5",
  "MFDFA>=0.4,<0.5",
  "pyts>=0.13,<0.14",
  "gatspy>=0.3,<0.4",
  # Utilities / IO / Viz
  "plotly>=6.5,<6.6",
  "tqdm>=4.67,<4.68",
  "PyYAML>=6.0,<6.1",
]
